[Speech SDK example] Fix resume logic for files that failed in the middle for whatever reason.
Fix resume logic for files that failed in the middle for files that failed in the middle of transcription for whatever reason. Beneficial if there were partial results output already so that we don't have to transcribe the file from scratch and can instead pick up from whatever offset where we left off.