[Speech SDK example] Add timeout in speech sdk to handle extremely rare cases where speech SDK freezes.
[Speech SDK example] Add timeout in speech sdk to handle extremely rare cases where speech SDK freezes.
Because the pool worker (and WorkItemProcessor) will just wait forever for a file to finish, if the file deadlocks because of a rare bug in Speech SDK, it will mean that the whole batch will get deadlocked. So here we don't allow the work item to get into this state. (Follow up work will add a generic heartbeat feature into batchkit for similiar optional functionality with arbitary work items).